split `ProducerController.cs` into two files
