+++ src/Aaron.Akka.ReliableDelivery/Aaron.Akka.ReliableDelivery.csproj
@@ -7,7 +7,7 @@
     </PropertyGroup>
 
     <ItemGroup>
-        <PackageReference Include="Akka.Streams"/>
+        <PackageReference Include="Akka.Streams" />
     </ItemGroup>
